+++ frontend/fingertips-frontend/lib/downloadHelpers/convertToCsv.ts
@@ -1,5 +1,3 @@
-import { PopulationDataForArea } from '../chartHelpers/preparePopulationData';
-
 const lineBreak = '\r\n';
 const delimiter = ',';
 
@@ -21,78 +19,3 @@ export const convertRowToCsv = (row: CsvRow): string =>
 
 export const convertToCsv = (data: CsvData): string =>
   data.map(convertRowToCsv).join(lineBreak).trim();
-
-export function populationPyramidTableToCsvArray({
-  indicatorId,
-  indicatorName,
-  period,
-  populationDataForArea,
-  populationDataForBenchmark,
-  populationDataForGroup,
-}: {
-  indicatorId: string;
-  indicatorName: string;
-  period: string;
-  populationDataForArea: PopulationDataForArea;
-  populationDataForBenchmark?: PopulationDataForArea;
-  populationDataForGroup?: PopulationDataForArea;
-}): CsvField[][] {
-  const header: CsvField[] = [
-    'Indicator ID',
-    'Indicator name',
-    'Period',
-    'Area code',
-    'Area name',
-    'Age range',
-    'Male',
-    'Female',
-  ];
-
-  const rows: CsvField[][] = [];
-  if (populationDataForBenchmark)
-    rows.push(
-      ...populationPyramidTableAreaToCsvArray(
-        indicatorId,
-        indicatorName,
-        period,
-        populationDataForBenchmark
-      )
-    );
-  if (populationDataForGroup)
-    rows.push(
-      ...populationPyramidTableAreaToCsvArray(
-        indicatorId,
-        indicatorName,
-        period,
-        populationDataForGroup
-      )
-    );
-  rows.push(
-    ...populationPyramidTableAreaToCsvArray(
-      indicatorId,
-      indicatorName,
-      period,
-      populationDataForArea
-    )
-  );
-
-  return [header, ...rows];
-}
-
-function populationPyramidTableAreaToCsvArray(
-  indicatorId: string,
-  indicatorName: string,
-  period: string,
-  populationDataForArea: PopulationDataForArea
-): CsvField[][] {
-  return populationDataForArea.ageCategories.map((ageCategory, i) => [
-    indicatorId,
-    indicatorName,
-    period,
-    populationDataForArea.areaCode,
-    populationDataForArea.areaName,
-    ageCategory,
-    populationDataForArea.maleSeries[i],
-    populationDataForArea.femaleSeries[i],
-  ]);
-}
